core: move the vm namespace upward
Because there's currently nothing else rather than the vm namespace.
