+++ src/eval/run/bitwise.rs
@@ -1,8 +1,8 @@
 //! Bitwise instructions
 
-use util::bigint::M256;
+use bigint::M256;
 
-use vm::Memory;
+use ::Memory;
 use super::State;
 
 pub fn iszero<M: Memory + Default>(state: &mut State<M>) {
