--- src/util/address.rs
@@ -1,2 +1,0 @@
-//! Ethereum address
-pub use bigint::Address;
