--- src/util/gas.rs
@@ -1,2 +1,0 @@
-//! Ethereum gas
-pub use bigint::Gas;
