+++ src/util/mod.rs
@@ -1,7 +1,2 @@
 //! Utilities of big integers, address, gas and opcodes
-pub mod bigint;
-pub mod address;
-pub mod gas;
 pub mod opcode;
-
-pub use hexutil::{read_hex, ParseHexError};
